requirements.txt
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
streamlit==1.15.2
openpyxl==3.0.10
